+++ app/src/main/java/com/example/geekhavencommunityapp/activities/OnboardingPage.kt
@@ -2,8 +2,6 @@ package com.example.geekhavencommunityapp
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.View
-import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
@@ -12,15 +10,12 @@ import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
-import com.example.geekhavencommunityapp.activities.BaseHomeActivity
+import com.example.geekhavencommunityapp.activities.signinActivity
 import com.example.geekhavencommunityapp.activities.usernameActivity
 import com.example.geekhavencommunityapp.fragments.Intro1
 import com.example.geekhavencommunityapp.fragments.Intro2
 import com.example.geekhavencommunityapp.fragments.Intro3
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator
-import com.tbuonomo.viewpagerdotsindicator.WormDotsIndicator
 
 class OnboardingPage : AppCompatActivity() {
 
@@ -50,7 +45,7 @@ class OnboardingPage : AppCompatActivity() {
         // skip button
         val skip: TextView = findViewById(R.id.btnSkip)
         skip.setOnClickListener {
-            val intent = Intent(this, usernameActivity::class.java)
+            val intent = Intent(this, signinActivity::class.java)
             startActivity(intent)
         }
 
@@ -58,7 +53,7 @@ class OnboardingPage : AppCompatActivity() {
         next.setOnClickListener {
             val viewPager = findViewById<ViewPager2>(R.id.view_pager)
             if (viewPager.currentItem == 2) {
-                val intent = Intent(this, usernameActivity::class.java)
+                val intent = Intent(this, signinActivity::class.java)
                 startActivity(intent)
             }
             viewPager.currentItem = viewPager.currentItem + 1
