--- app/src/main/java/com/example/geekhavencommunityapp/activities/passwordActivity.kt
@@ -1,59 +1,0 @@
-package com.example.geekhavencommunityapp.activities
-
-import android.content.ContentValues.TAG
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.util.Log
-import android.widget.Button
-import android.widget.EditText
-import android.widget.Toast
-import androidx.activity.viewModels
-import com.example.geekhavencommunityapp.R
-import com.example.geekhavencommunityapp.UserModel
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.UserProfileChangeRequest
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.auth.userProfileChangeRequest
-import com.google.firebase.ktx.Firebase
-
-class passwordActivity : AppCompatActivity() {
-    private lateinit var auth : FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_password)
-
-
-        auth = Firebase.auth
-
-        val next = findViewById<Button>(R.id.log_in)
-        val password = findViewById<EditText>(R.id.password)
-
-        next.setOnClickListener{
-
-            val currentIntent = intent
-
-
-            auth.createUserWithEmailAndPassword(UserModel.email!! , password.text.toString())
-                .addOnCompleteListener(this) { task ->
-                if (task.isSuccessful) {
-                    // Sign in success, update UI with the signed-in user's information
-                    Log.d(TAG, "createUserWithEmail:success")
-                    val user = auth.currentUser
-                    startActivity(Intent(this, BaseHomeActivity::class.java))
-
-                } else {
-                    // If sign in fails, display a message to the user.
-                    Log.w(TAG, "createUserWithEmail:failure", task.exception)
-                    Toast.makeText(
-                        baseContext,
-                        "Authentication failed.",
-                        Toast.LENGTH_SHORT,
-                    ).show()
-                }
-            }
-
-        }
-    }
-}
