--- app/src/main/res/color/bottom_navigation_colors.xml
@@ -1,4 +1,0 @@
-<resources>
-    <color name="bottom_nav_icon_color">#FFFFFF</color>
-    <color name="bottom_nav_text_color">#FFFFFF</color>
-</resources>
