+++ app/src/main/res/drawable/circle.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <solid android:color="#0a0c16"/>
+    <size android:width="80dp" android:height="80dp"
+        />
+    <stroke android:color="#969595" android:width="2dp"/>
+
+</shape>
