+++ app/src/main/res/layout/activity_login_screen.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.emailActivity">
+    tools:context=".activities.signinActivity">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
