+++ app/src/main/res/values/colors.xml
@@ -12,5 +12,8 @@
     <color name="green">#22EA2A</color>
     <color name="blue">#5F93E1</color>
     <color name="grey">#B2BEB5</color>
+    <color name="bottom_nav_icon_color">#FFFFFF</color>
+    <color name="bottom_nav_text_color">#FFFFFF</color>
+
 
 </resources>
