+++ app/src/main/res/values/styles.xml
@@ -13,4 +13,11 @@
         <item name="android:textSize">10sp</item>
     </style>
 
+    <style name="CustomTextInputLayout" parent="Widget.MaterialComponents.TextInputLayout.OutlinedBox">
+        <item name="boxStrokeColor">#B42B2B</item>
+    </style>
+
+
+
+
 </resources>
